requirements.txt
langchain>=0.3.7
langchain-community>=0.3.7
langgraph>=0.2.45
langchain-core>=0.3.15
langchain-chroma>=0.1.4
langchain-huggingface>=0.1.0
langchain-ollama>=0.2.0
langchain-google-genai>=2.0.4
langchain-openai>=0.2.6
transformers>=4.46.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
tenacity>=8.2.0
termcolor>=2.4.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
